+++ src/smsfusion/__init__.py
@@ -1,21 +1,22 @@
-from . import _ins_v2  # hidden
 from . import benchmark, calibrate, constants, noise
 from ._coning_sculling import ConingScullingAlg
-from ._ins import AHRS, VRU, AidedINS, FixedNED, StrapdownINS, gravity
+
+# from ._ins import _ahrs  # hidden
+from ._ins import AHRS, VRU, AHRSv2, AidedINS, FixedNED, StrapdownINS, VRUv2, gravity
 from ._smoothing import FixedIntervalSmoother
 from ._transforms import quaternion_from_euler
 
 __all__ = [
     "AHRS",
-    "AidedINS",
+    # "AidedINS",
     "benchmark",
     "constants",
     "calibrate",
     "FixedIntervalSmoother",
-    "FixedNED",
-    "gravity",
+    # "FixedNED",
+    # "gravity",
     "noise",
-    "StrapdownINS",
+    # "StrapdownINS",
     "VRU",
     "quaternion_from_euler",
     "ConingScullingAlg",
