+++ src/smsfusion/_ins/__init__.py
@@ -1,0 +1,3 @@
+from ._ahrs import AHRS as AHRSv2
+from ._ains import AHRS, VRU, AidedINS, FixedNED, StrapdownINS, gravity
+from ._vru import VRU as VRUv2
